+++ playwright/core/types.ts
@@ -1,3 +1,5 @@
+import type React from 'react';
+
 import type {MountOptions, MountResult} from '@playwright/experimental-ct-react';
 import type {
     Locator,
@@ -11,8 +13,8 @@ import type {
 
 interface ComponentFixtures {
     mount<HooksConfig>(
-        component: JSX.Element,
-        options?: MountOptions<HooksConfig>,
+        component: React.JSX.Element,
+        options?: MountOptions<HooksConfig> & {width?: number},
     ): Promise<MountResult>;
 }
 
@@ -32,7 +34,8 @@ export interface ExpectScreenshotFixture {
     (props?: CaptureScreenshotParams): Promise<void>;
 }
 
-interface CaptureScreenshotParams extends PageScreenshotOptions {
+export interface CaptureScreenshotParams extends PageScreenshotOptions {
     nameSuffix?: string;
     component?: Locator;
+    themes?: Array<'light' | 'dark'>;
 }