+++ src/components/AvatarStack/AvatarStack.tsx
@@ -1,73 +1,80 @@
 import React from 'react';
 
-import {Avatar} from '../Avatar';
 import {block} from '../utils/cn';
 
 import {AvatarStackItem} from './AvatarStackItem';
+import {AvatarStackMore} from './AvatarStackMore';
 import {AvatarStackMoreButton} from './AvatarStackMoreButton';
-import i18n from './i18n';
+import {AVATAR_STACK_DEFAULT_MAX} from './constants';
 import type {AvatarStackProps} from './types';
 
 import './AvatarStack.scss';
 
 const b = block('avatar-stack');
 
-const AvatarStackComponent = ({
-    max = 3,
-    total,
-    overlapSize = 's',
-    size,
-    children,
-    className,
-    renderMore,
-}: AvatarStackProps) => {
-    const visibleItems: React.ReactElement[] = [];
-    let moreItems = 0;
-    const normalizedMax = max < 1 ? 1 : max;
+const AvatarStackComponent = React.forwardRef<HTMLUListElement, AvatarStackProps>(
+    (
+        {
+            max = AVATAR_STACK_DEFAULT_MAX,
+            total,
+            overlapSize = 's',
+            size,
+            children,
+            className,
+            renderMore,
+        },
+        ref,
+    ) => {
+        const visibleItems: React.ReactElement[] = [];
 
-    React.Children.forEach(children, (child) => {
-        if (!React.isValidElement(child)) {
-            return;
-        }
+        /** All avatars amount */
+        const normalizedTotal = total ? Math.max(total, max) : React.Children.count(children);
 
-        const item = <AvatarStackItem key={visibleItems.length}>{child}</AvatarStackItem>;
+        /** Amount avatars to be visible (doesn't include badge with remaining avatars) */
+        let normalizedMax = max < 1 ? 1 : max;
+        // Skip rendering badge with +1, just show avatar instead
+        normalizedMax = normalizedTotal - normalizedMax > 1 ? normalizedMax : normalizedTotal;
 
-        if (visibleItems.length <= normalizedMax) {
-            visibleItems.unshift(item);
-        } else {
-            moreItems += 1;
-        }
-    });
+        /** Remaining avatars */
+        const moreItems = normalizedTotal - normalizedMax;
 
-    moreItems = Math.max(moreItems, total ? total - normalizedMax : 0);
+        React.Children.forEach(children, (child) => {
+            if (!React.isValidElement(child)) {
+                return;
+            }
 
-    const hasMoreButton = moreItems > 0;
-    /** Avatars + more button, or just avatars, when avatars count is equal to `max` or less */
-    const normalOverflow = moreItems >= 1;
+            const item = <AvatarStackItem key={visibleItems.length}>{child}</AvatarStackItem>;
 
-    return (
-        // Safari remove role=list with some styles, applied to li items, so we need
-        // to restore role manually
-        // eslint-disable-next-line jsx-a11y/no-redundant-roles
-        <ul className={b({'overlap-size': overlapSize}, className)} role={'list'}>
-            {hasMoreButton ? (
-                <AvatarStackItem key="more-button">
-                    {renderMore ? (
-                        renderMore({count: moreItems})
-                    ) : (
-                        <Avatar
-                            text={`+${moreItems}`}
-                            aria-label={i18n('more', {count: moreItems})}
-                            size={size}
-                        />
-                    )}
-                </AvatarStackItem>
-            ) : null}
-            {normalOverflow ? visibleItems.slice(0, normalizedMax) : visibleItems}
-        </ul>
-    );
-};
+            if (visibleItems.length < normalizedMax) {
+                visibleItems.unshift(item);
+            }
+        });
+
+        const hasMoreButton = moreItems > 0;
+
+        return (
+            // Safari remove role=list with some styles, applied to li items, so we need
+            // to restore role manually
+            // eslint-disable-next-line jsx-a11y/no-redundant-roles
+            <ul className={b({'overlap-size': overlapSize}, className)} role={'list'} ref={ref}>
+                {hasMoreButton ? (
+                    <AvatarStackItem key="more-button">
+                        {renderMore ? (
+                            renderMore({count: moreItems})
+                        ) : (
+                            <AvatarStackMore count={moreItems} size={size} />
+                        )}
+                    </AvatarStackItem>
+                ) : null}
+                {visibleItems}
+            </ul>
+        );
+    },
+);
 
 AvatarStackComponent.displayName = 'AvatarStack';
 
-export const AvatarStack = Object.assign(AvatarStackComponent, {MoreButton: AvatarStackMoreButton});
+export const AvatarStack = Object.assign(AvatarStackComponent, {
+    More: AvatarStackMore,
+    MoreButton: AvatarStackMoreButton,
+});