+++ src/components/Slider/SliderTooltip/SliderTooltip.tsx
@@ -12,35 +12,20 @@ import './SliderTooltip.scss';
 const b = block('slider-tooltip');
 
 type SliderTooltipProps = {
-    value: number;
     className?: string;
     style?: React.CSSProperties;
     stateModifiers: Omit<StateModifiers, 'hasTooltip'>;
-};
+} & React.PropsWithChildren;
 
-export const SliderTooltip = ({value, className, style, stateModifiers}: SliderTooltipProps) => {
-    const preventAction = React.useCallback((e: React.SyntheticEvent) => {
-        //make tooltip non-interactive
-        e.preventDefault();
-        e.stopPropagation();
-    }, []);
-    return (
-        // eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions
-        <div
-            className={b(stateModifiers, className)}
-            style={style}
-            onClick={preventAction}
-            onMouseDown={preventAction}
-            onTouchStart={preventAction}
-        >
-            <div className={b('card', stateModifiers)}>
-                {value}
-                {stateModifiers.disabled && (
-                    //use this element to prevent crossing effect
-                    <SliderTooltipPin className={b('pin', {background: true})} />
-                )}
-                <SliderTooltipPin className={b('pin')} />
-            </div>
+export const SliderTooltip = ({className, style, stateModifiers, children}: SliderTooltipProps) => (
+    <div className={b(stateModifiers, className)} style={style}>
+        <div className={b('card', stateModifiers)}>
+            {children}
+            {stateModifiers.disabled && (
+                //use this element to prevent crossing effect
+                <SliderTooltipPin className={b('pin', {background: true})} />
+            )}
+            <SliderTooltipPin className={b('pin')} />
         </div>
-    );
-};
+    </div>
+);