+++ src/components/Toaster/types.ts
@@ -1,11 +1,13 @@
 import type {ButtonView} from '../Button';
 
+import type {TOAST_THEMES} from './constants';
+
 export type ToasterArgs = {
     className?: string;
     mobile?: boolean;
 };
 
-export type ToastTheme = 'normal' | 'info' | 'success' | 'warning' | 'danger' | 'utility';
+export type ToastTheme = (typeof TOAST_THEMES)[number];
 
 export type ToastAction = {
     onClick: VoidFunction;
@@ -23,6 +25,8 @@ export type ToastProps = {
     theme?: ToastTheme;
     isClosable?: boolean;
     actions?: ToastAction[];
+
+    onClose?: () => void;
 
     /** Function. Use for toast icon customization. By default type-based behavior is used */
     renderIcon?: (toastProps: ToastProps) => React.ReactNode;