+++ src/hooks/useBodyScrollLock/useBodyScrollLock.ts
@@ -1,5 +1,28 @@
 import React from 'react';
 
+const PROPERTY_PADDING_RIGHT = 'padding-right';
+const PROPERTY_PADDING_BOTTOM = 'padding-bottom';
+const PROPERTY_OVERFLOW = 'overflow';
+
+const STORED_BODY_STYLE_KEYS = [
+    PROPERTY_OVERFLOW,
+    PROPERTY_PADDING_RIGHT,
+    PROPERTY_PADDING_BOTTOM,
+] as const;
+
+type StoredBodyStyleKeys = (typeof STORED_BODY_STYLE_KEYS)[number];
+type StoredBodyStyle = Partial<Record<StoredBodyStyleKeys, string>>;
+
+function getStoredStyles(): StoredBodyStyle {
+    const styles: StoredBodyStyle = {};
+
+    for (const property of STORED_BODY_STYLE_KEYS) {
+        styles[property] = document.body.style.getPropertyValue(property);
+    }
+
+    return styles;
+}
+
 export interface UseBodyScrollLockProps {
     enabled: boolean;
 }
@@ -7,7 +30,7 @@ export interface UseBodyScrollLockProps {
 export type BodyScrollLockProps = UseBodyScrollLockProps;
 
 let locks = 0;
-let storedBodyStyle: string | undefined;
+let storedBodyStyle: StoredBodyStyle = {};
 
 export function useBodyScrollLock({enabled}: UseBodyScrollLockProps) {
     React.useLayoutEffect(() => {
@@ -35,22 +58,32 @@ function setBodyStyles() {
     const xScrollbarWidth = getXScrollbarWidth();
     const bodyPadding = getBodyComputedPadding();
 
-    storedBodyStyle = document.body.style.cssText;
-    document.body.style.overflow = 'hidden';
+    storedBodyStyle = getStoredStyles();
+
+    document.body.style.setProperty(PROPERTY_OVERFLOW, 'hidden');
 
     if (yScrollbarWidth) {
-        document.body.style.paddingRight = `${bodyPadding.right + yScrollbarWidth}px`;
+        document.body.style.setProperty(
+            PROPERTY_PADDING_RIGHT,
+            `${bodyPadding.right + yScrollbarWidth}px`,
+        );
     }
     if (xScrollbarWidth) {
-        document.body.style.paddingBottom = `${bodyPadding.bottom + xScrollbarWidth}px`;
+        document.body.style.setProperty(
+            PROPERTY_PADDING_BOTTOM,
+            `${bodyPadding.bottom + xScrollbarWidth}px`,
+        );
     }
 }
 
 function restoreBodyStyles() {
-    if (storedBodyStyle) {
-        document.body.style.cssText = storedBodyStyle;
-    } else {
-        document.body.removeAttribute('style');
+    for (const property of STORED_BODY_STYLE_KEYS) {
+        const storedProperty = storedBodyStyle[property];
+        if (storedProperty) {
+            document.body.style.setProperty(property, storedProperty);
+        } else {
+            document.body.style.removeProperty(property);
+        }
     }
 }