+++ CHANGELOG.md
@@ -1,5 +1,124 @@
 # Changelog
 
+## [6.35.2](https://github.com/gravity-ui/uikit/compare/v6.35.1...v6.35.2) (2024-11-14)
+
+
+### Bug Fixes
+
+* **PlaceholderContainer:** remove href attribute fallback in actions ([#1947](https://github.com/gravity-ui/uikit/issues/1947)) ([60a8b92](https://github.com/gravity-ui/uikit/commit/60a8b9240b51d78f0aa5e711a6bed8e30e049814))
+* **useBodyScrollLock:** correctly restore styles ([#1950](https://github.com/gravity-ui/uikit/issues/1950)) ([b41a53a](https://github.com/gravity-ui/uikit/commit/b41a53a477d479b925331bbc227862942da04d00))
+
+## [6.35.1](https://github.com/gravity-ui/uikit/compare/v6.35.0...v6.35.1) (2024-11-13)
+
+
+### Bug Fixes
+
+* fix incorrect scss "use" ([#1943](https://github.com/gravity-ui/uikit/issues/1943)) ([c9506dd](https://github.com/gravity-ui/uikit/commit/c9506dd5cfb78f155ea4e38f09e060233c4f691b))
+
+## [6.35.0](https://github.com/gravity-ui/uikit/compare/v6.34.0...v6.35.0) (2024-11-11)
+
+
+### Features
+
+* **FilePreview:** add component ([#1880](https://github.com/gravity-ui/uikit/issues/1880)) ([b5de671](https://github.com/gravity-ui/uikit/commit/b5de67126a76467853957df6d8e5436cc460886e))
+* **NumberInput:** add new component ([#1826](https://github.com/gravity-ui/uikit/issues/1826)) ([75be05e](https://github.com/gravity-ui/uikit/commit/75be05e42c95226592cdd2139894e9ff791280af))
+* **PlaceholderContainer:** add qa prop ([#1925](https://github.com/gravity-ui/uikit/issues/1925)) ([77ad28d](https://github.com/gravity-ui/uikit/commit/77ad28d8aa0af910999e0ed54fce806767241946))
+
+
+### Bug Fixes
+
+* **PlaceholderContainer:** add component to index file ([#1924](https://github.com/gravity-ui/uikit/issues/1924)) ([cdce171](https://github.com/gravity-ui/uikit/commit/cdce1715a6fe5ceebff7b43c9146c3c1bef9f093))
+* **Popover:** use visible title for popup a11y label ([#1941](https://github.com/gravity-ui/uikit/issues/1941)) ([f9d75e8](https://github.com/gravity-ui/uikit/commit/f9d75e8b1d902fd526b4a089d7bf6675299f2584))
+* **Toaster:** reveal animation lag ([#1927](https://github.com/gravity-ui/uikit/issues/1927)) ([e16fa1c](https://github.com/gravity-ui/uikit/commit/e16fa1ccc7e1f053f4ff696d24ee6d4bbfd35a6a))
+
+## [6.34.0](https://github.com/gravity-ui/uikit/compare/v6.33.0...v6.34.0) (2024-10-25)
+
+
+### Features
+
+* **ListItemExandIcon:** added new list component for expanded icon view ([#1762](https://github.com/gravity-ui/uikit/issues/1762)) ([b6516f4](https://github.com/gravity-ui/uikit/commit/b6516f46ce157a1297a5887e25487f621da07b43))
+
+
+### Bug Fixes
+
+* react 19 compatibility ([#1916](https://github.com/gravity-ui/uikit/issues/1916)) ([29f2558](https://github.com/gravity-ui/uikit/commit/29f255803582d9b6791f11b080b80d36626dd03b))
+* **styles:** rules order in block ([#1912](https://github.com/gravity-ui/uikit/issues/1912)) ([2d9c88e](https://github.com/gravity-ui/uikit/commit/2d9c88ee6a7040d33369c1a47d2022ec9f129d7c))
+
+## [6.33.0](https://github.com/gravity-ui/uikit/compare/v6.32.0...v6.33.0) (2024-10-22)
+
+
+### Features
+
+* **Dialog:** add className property to DialogFooter ([#1909](https://github.com/gravity-ui/uikit/issues/1909)) ([679c537](https://github.com/gravity-ui/uikit/commit/679c537a0a34add7e6dca4a3ba649934a68342e6))
+* **Table:** add width property ([#1907](https://github.com/gravity-ui/uikit/issues/1907)) ([7eb5e1e](https://github.com/gravity-ui/uikit/commit/7eb5e1e38cb9e79afaf988480b1b2d5dc998ff04))
+* **Toaster:** added onClose callback ([#1902](https://github.com/gravity-ui/uikit/issues/1902)) ([3069194](https://github.com/gravity-ui/uikit/commit/30691949ccce25e54614b5a0b49c7b11f7abf3e9))
+
+
+### Bug Fixes
+
+* **DefinitionList:** layout fixes ([#1910](https://github.com/gravity-ui/uikit/issues/1910)) ([e5cde3a](https://github.com/gravity-ui/uikit/commit/e5cde3ad58cfb1d8436c2a4ad75e074dfb94a741))
+
+## [6.32.0](https://github.com/gravity-ui/uikit/compare/v6.31.0...v6.32.0) (2024-10-17)
+
+
+### Features
+
+* add DefinitionList and HelpMark ([#1731](https://github.com/gravity-ui/uikit/issues/1731)) ([c587116](https://github.com/gravity-ui/uikit/commit/c587116b40e79c1ad7a2427f73d031b6efa4109c))
+* **AvatarStack:** forward ref to root element of component ([#1904](https://github.com/gravity-ui/uikit/issues/1904)) ([469d935](https://github.com/gravity-ui/uikit/commit/469d93586d6bb82b93a865541d8a974caa67d3b1))
+
+
+### Bug Fixes
+
+* **ListItemView:** fix type ListItemViewProps ([#1897](https://github.com/gravity-ui/uikit/issues/1897)) ([e49779a](https://github.com/gravity-ui/uikit/commit/e49779a4e519a6e89db3cecfe12ea66e0c720d3f))
+* **Toaster:** change spacing before actions ([#1891](https://github.com/gravity-ui/uikit/issues/1891)) ([235ea21](https://github.com/gravity-ui/uikit/commit/235ea21a548d39b6ebf22e31d9037abc83eb9fcd))
+* **types:** fix i18n types ([#1898](https://github.com/gravity-ui/uikit/issues/1898)) ([21954db](https://github.com/gravity-ui/uikit/commit/21954db04c9edc5d5068007fd4f137e0732e2b47))
+* **useBodyScrollLock:** avoid settings unrelated styles ([#1832](https://github.com/gravity-ui/uikit/issues/1832)) ([9d7fd36](https://github.com/gravity-ui/uikit/commit/9d7fd3649c7c82a1c3d68af7d2f8ceffe7172eef))
+
+## [6.31.0](https://github.com/gravity-ui/uikit/compare/v6.30.1...v6.31.0) (2024-10-09)
+
+
+### Features
+
+* **TreeSelect:** added error state view and hasClear prop ([#1885](https://github.com/gravity-ui/uikit/issues/1885)) ([9849e4b](https://github.com/gravity-ui/uikit/commit/9849e4bacc6b11bdfef191e31dde28a75a5e723c))
+
+
+### Bug Fixes
+
+* **Select:** revert [#1694](https://github.com/gravity-ui/uikit/issues/1694) changes ([#1887](https://github.com/gravity-ui/uikit/issues/1887)) ([523caf8](https://github.com/gravity-ui/uikit/commit/523caf8915a5ee7cbb0652886f2ee247baeabb2e))
+
+## [6.30.1](https://github.com/gravity-ui/uikit/compare/v6.30.0...v6.30.1) (2024-10-04)
+
+
+### Bug Fixes
+
+* **AvatarStack:** remaining avatars calculation ([#1883](https://github.com/gravity-ui/uikit/issues/1883)) ([76943c9](https://github.com/gravity-ui/uikit/commit/76943c9174658be0da60d98ce50c290ef2ca45a1))
+
+## [6.30.0](https://github.com/gravity-ui/uikit/compare/v6.29.0...v6.30.0) (2024-10-03)
+
+
+### Features
+
+* **ActionsPanel:** add component and showcases ([#1873](https://github.com/gravity-ui/uikit/issues/1873)) ([5ce703f](https://github.com/gravity-ui/uikit/commit/5ce703f9b0da81493262035e8259d0f0aedc3552))
+* **ClipboardButton:** delay tooltip's closing animation after copying ([#1735](https://github.com/gravity-ui/uikit/issues/1735)) ([20f19dd](https://github.com/gravity-ui/uikit/commit/20f19ddca63dca4d34b9fd62b4e0b4073fdcd2be))
+
+
+### Bug Fixes
+
+* **AvatarStack:** correctly display more than 9 remaining avatars ([#1882](https://github.com/gravity-ui/uikit/issues/1882)) ([0abd7f1](https://github.com/gravity-ui/uikit/commit/0abd7f1bc4f01a5616d4cce532e9f9a6cc4f5c53))
+
+## [6.29.0](https://github.com/gravity-ui/uikit/compare/v6.28.0...v6.29.0) (2024-09-27)
+
+
+### Features
+
+* **Slider:** redesign API to get more control on marks ([#1744](https://github.com/gravity-ui/uikit/issues/1744)) ([2b5c26e](https://github.com/gravity-ui/uikit/commit/2b5c26e6144cc74bd2467ae6f5fe9e6dd3a14197))
+* **Table:** add dynamic href support for row actions in withTableActions ([#1874](https://github.com/gravity-ui/uikit/issues/1874)) ([17b9a66](https://github.com/gravity-ui/uikit/commit/17b9a663fea91515b1fa646ddb2003f4e70016a5))
+
+
+### Bug Fixes
+
+* **Select:** fix tick icon color ([#1878](https://github.com/gravity-ui/uikit/issues/1878)) ([fc82360](https://github.com/gravity-ui/uikit/commit/fc82360a74ae02ce1b22e1861befb8a891679f15))
+
 ## [6.28.0](https://github.com/gravity-ui/uikit/compare/v6.27.2...v6.28.0) (2024-09-24)